Move log_raiden_version test to separate file
And also fix a linter error.
